Optimisation of Feed Forward Layer : added sse2 implementation for FCL forward and backpropagation and for matrix operations : Add, *=

+++ src/InitFunc.cpp
@@ -3,34 +3,34 @@
 
 void XavierInit(const int inputSize, Matrix* weights)
 {
-    double upper = 1.0 / sqrt((double) inputSize);
-    double lower = -upper;
+    float upper = 1.0 / sqrt((float) inputSize);
+    float lower = -upper;
 
     for (int i = 0; i < weights->size(); i++)
     {
-        weights[0][i] = lower + (rand() / ((double) RAND_MAX) * (upper - (lower)));
+        weights[0][i] = lower + (rand() / ((float) RAND_MAX) * (upper - (lower)));
     }
 };
 
 
 void NormalizedXavierInit(const int inputSize, const int outputSize, Matrix* weights)
 {
-    double upper = (sqrt(6.0) / sqrt((double) inputSize + (double) outputSize));
-    double lower = -upper;
+    float upper = (sqrt(6.0) / sqrt((float) inputSize + (float) outputSize));
+    float lower = -upper;
 
     for (int i = 0; i < weights->size(); i++)
     {
-        weights[0][i] = lower + (rand() / ((double) RAND_MAX) * (upper - (lower)));
+        weights[0][i] = lower + (rand() / ((float) RAND_MAX) * (upper - (lower)));
     }
 };
 
 void HeInit(const int inputSize, Matrix* weights)
 {
-    double range = sqrt(2.0 / (double) inputSize);
+    float range = sqrt(2.0 / (float) inputSize);
 
     for (int i = 0; i < weights->size(); i++)
     {
-        weights[0][i] = (rand() / ((double) RAND_MAX) - 0.5) * 2 * range;
+        weights[0][i] = (rand() / ((float) RAND_MAX) - 0.5) * 2 * range;
     }
 
 };
